memory: increate the amount of memory
So far, we only define the maximum amount of memory for a VM. The
allocation was still at 768MB by default.
 "If VIR_DOMAIN_MEM_MAXIMUM is set, the change affects domain's maximum
memory size rather than current memory size."

With this change, we ensure the VM get the amount of memory that we
expect.

See: https://libvirt.org/html/libvirt-libvirt-domain.html#virDomainSetMemoryFlags